Fixed bug in Microsoft.Application.targets with extensionless files
There is a following bug in Microsoft.Application.targets (the same in its origin - Microsoft.Web.targets):

If you add a file without extension (e.q. "a", "b" instead of "a.cs", "b.txt") as "Content" to your project in VS - it will not be copied to _PublishedPallication. This is a fix for the bug.